Add verification to FormatMessage in readlog.c
Fixes problem when messages from the event log references non-existent
formatting arguments, or "inserts" as FormatMessage calls them.

An example of an insert is "%1" in a message "An exception occured in %1.",
where the "args" argument list argument to FormatMessage should contain
the insert (substitution) value of "%1" at index 0. This causes memory
access violations when the message contains e.g. "%10", and args only
has 3 elements.

The behavior is changed, so the message is scanned for its highest
number argument. If it is not equal to the argument list length, it
does not try to perform any inserts, and instead reports the message
without modifications. In other words, the final message still contains
formatting sequences such as "%2".

Messages with correct inserts have their inserts applied just as before.